+++ src/app/modules/search-parameters/search-parameters.component.spec.ts
@@ -7,7 +7,6 @@ import {
 } from '../../shared/fhir-backend/fhir-backend.service';
 import { configureTestingModule } from 'src/test/helpers';
 import { ErrorManager } from '../../shared/error-manager/error-manager.service';
-import { RouterTestingModule } from '@angular/router/testing';
 
 describe('SearchParametersComponent', () => {
   let component: SearchParametersComponent;
@@ -17,7 +16,7 @@ describe('SearchParametersComponent', () => {
   beforeEach(async () => {
     await configureTestingModule({
       declarations: [SearchParametersComponent],
-      imports: [SearchParametersModule, RouterTestingModule],
+      imports: [SearchParametersModule],
       providers: [ErrorManager]
     });
     fhirBackend = TestBed.inject(FhirBackendService);